Changed zip, map, filter to return zip, map, filter objects in python 3
