+++ src/main.js
@@ -47,6 +47,9 @@ require("./file.js");
 require("./ffi.js");
 require("./iterator.js");
 require("./enumerate.js");
+require("./filter.js");
+require("./zip.js");
+require("./map.js");
 require("./token.js");
 require("./tokenize.js");
 require("../gen/parse_tables.js");
